Squashed commit of the following:

commit acaadd21e4b1ac67d2e2ce16bfbc1ed0883bf1df
Author: DEXREQ Automation <noreply+dexreq@oracle.com>
Date:   Fri May 15 19:57:33 2020 +0000

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 15 May 2020 19:57:33 +0000
    Running generation for: [[DEXREQ-1284, DEXREQ-1278, DEXREQ-1277, DEXREQ-1283, DEXREQ-1274, DEXREQ-1269, DEXREQ-1285, DEXREQ-1286, DEXREQ-1280]]

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 15 May 2020 19:18:49 +0000
    Updating pom.xml for [[DEXREQ-1284, DEXREQ-1278, DEXREQ-1277, DEXREQ-1283, DEXREQ-1274, DEXREQ-1269, DEXREQ-1285, DEXREQ-1286, DEXREQ-1280]]

    (automatically combined)

+++ dataintegration/task_run.go
@@ -42,19 +42,19 @@ type TaskRun struct {
 	Status TaskRunStatusEnum `mandatory:"false" json:"status,omitempty"`
 
 	// startTimeMillis
-	StartTimeMillis *int `mandatory:"false" json:"startTimeMillis"`
+	StartTimeMillis *int64 `mandatory:"false" json:"startTimeMillis"`
 
 	// endTimeMillis
-	EndTimeMillis *int `mandatory:"false" json:"endTimeMillis"`
+	EndTimeMillis *int64 `mandatory:"false" json:"endTimeMillis"`
 
 	// lastUpdated
-	LastUpdated *int `mandatory:"false" json:"lastUpdated"`
+	LastUpdated *int64 `mandatory:"false" json:"lastUpdated"`
 
 	// Number of records processed in task run.
-	RecordsWritten *int `mandatory:"false" json:"recordsWritten"`
+	RecordsWritten *int64 `mandatory:"false" json:"recordsWritten"`
 
 	// Number of bytes processed in task run.
-	BytesProcessed *int `mandatory:"false" json:"bytesProcessed"`
+	BytesProcessed *int64 `mandatory:"false" json:"bytesProcessed"`
 
 	// Error message if status is ERROR
 	ErrorMessage *string `mandatory:"false" json:"errorMessage"`
