+++ dataintegration/task_run_details.go
@@ -40,19 +40,19 @@ type TaskRunDetails struct {
 	Status TaskRunDetailsStatusEnum `mandatory:"false" json:"status,omitempty"`
 
 	// startTimeMillis
-	StartTimeMillis *int `mandatory:"false" json:"startTimeMillis"`
+	StartTimeMillis *int64 `mandatory:"false" json:"startTimeMillis"`
 
 	// endTimeMillis
-	EndTimeMillis *int `mandatory:"false" json:"endTimeMillis"`
+	EndTimeMillis *int64 `mandatory:"false" json:"endTimeMillis"`
 
 	// lastUpdated
-	LastUpdated *int `mandatory:"false" json:"lastUpdated"`
+	LastUpdated *int64 `mandatory:"false" json:"lastUpdated"`
 
 	// Number of records processed in task run.
-	RecordsWritten *int `mandatory:"false" json:"recordsWritten"`
+	RecordsWritten *int64 `mandatory:"false" json:"recordsWritten"`
 
 	// Number of bytes processed in task run.
-	BytesProcessed *int `mandatory:"false" json:"bytesProcessed"`
+	BytesProcessed *int64 `mandatory:"false" json:"bytesProcessed"`
 
 	// The status of an object that can be set to value 1 for shallow references across objects, other values reserved.
 	ObjectStatus *int `mandatory:"false" json:"objectStatus"`
